EID-1297 Consolidate URLs to shared
The URLs can be used between apps, and only need changing in one place.
