+++ eidas-saml-parser/src/main/java/uk/gov/ida/notification/eidassaml/EidasSamlResource.java
@@ -7,6 +7,7 @@ import org.opensaml.saml.saml2.metadata.SPSSODescriptor;
 import se.litsec.opensaml.utils.ObjectUtils;
 import uk.gov.ida.notification.eidassaml.saml.validation.EidasAuthnRequestValidator;
 import uk.gov.ida.saml.security.validators.signature.SamlRequestSignatureValidator;
+import uk.gov.ida.notification.shared.Urls;
 
 import javax.validation.Valid;
 import javax.ws.rs.POST;
@@ -16,7 +17,7 @@ import javax.ws.rs.core.MediaType;
 import java.io.ByteArrayInputStream;
 import java.util.Base64;
 
-@Path("/eidasAuthnRequest")
+@Path(Urls.EidasSamlParserUrls.EIDAS_AUTHN_REQUEST_PATH)
 @Produces(MediaType.APPLICATION_JSON)
 public class EidasSamlResource {
