--- proxy-node-gateway/src/main/java/uk/gov/ida/notification/Urls.java
@@ -1,8 +1,0 @@
-package uk.gov.ida.notification;
-
-public interface Urls {
-
-    interface EidasSamlParserUrls {
-        String EIDAS_AUTHN_REQUEST_PATH = "/eidasAuthnRequest";
-    }
-}
