+++ proxy-node-gateway/src/main/java/uk/gov/ida/notification/resources/HubResponseResource.java
@@ -10,6 +10,7 @@ import uk.gov.ida.notification.exceptions.hubresponse.InvalidHubResponseExceptio
 import uk.gov.ida.notification.saml.HubResponseContainer;
 import uk.gov.ida.notification.saml.SamlFormMessageType;
 import uk.gov.ida.notification.saml.validation.HubResponseValidator;
+import uk.gov.ida.notification.shared.Urls;
 
 import javax.servlet.http.HttpSession;
 import javax.ws.rs.Consumes;
