--- proxy-node-translator/src/main/java/uk/gov/ida/notification/translator/Urls.java
@@ -1,9 +1,0 @@
-package uk.gov.ida.notification.translator;
-
-public interface Urls {
-
-    interface TranslatorUrls {
-        String TRANSLATOR_ROOT = "/translator";
-        String TRANSLATE_HUB_RESPONSE_PATH = "/translate-hub-response";
-    }
-}
